+++ packages/terra-data-grid/src/terra-dev-site/doc/data-grid/Examples/DatagridWithSubsections.2.doc.jsx
@@ -90,7 +90,7 @@ class SubsectionDataGrid extends React.Component {
       id: sectionId,
       text: sectionName,
       endAccessory: (sectionId === 'section_1') ? <span><Button text="Button 1" data-accessible-data-grid-content /><Button text="Button 2" data-accessible-data-grid-content /></span> : null,
-      isCollapsible: sectionId === 'section_2',
+      isCollapsible: sectionId === 'section_0',
       isCollapsed: this.state.collapsedSectionId === sectionId,
       rows: SubsectionDataGrid.buildRows(sectionId, numberOfRows),
     };
