+++ packages/terra-data-grid/src/utils/keycodes.js
@@ -1,0 +1,9 @@
+const KEYCODES = {
+  ENTER: 13,
+  SPACE: 32,
+  TAB: 9,
+  SHIFT: 16,
+};
+
+export default KEYCODES;
+export { KEYCODES };
